dt-demo-api/pyproject.toml
[project]
name = "dt-demo-api"
version = "0.1.0.dev0"
description = "Root API for the Hospital DT demo."
readme = "README.md"
authors = [
    { name = "Yin-Chi Chan", email = "ycc39@cam.ac.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "dt-demo-auth",
    "fastapi[standard]>=0.116.1",
    "jwt-pydantic>=0.0.7",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
    "uuid6>=2025.0.1",
]

[project.scripts]
main = "dt_demo_api:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
dt-demo-auth = { workspace = true }
